+++ src/arrayFilterTuple/arrayFilterTuple.ts
@@ -1,4 +1,5 @@
 import arrayFilter from "../arrayFilter/arrayFilter";
+import objectFreeze from "../objectFreeze";
 import ArrayFilterTupleFunction from "./ArrayFilterTupleFunction";
 
 /**
@@ -8,9 +9,10 @@ import ArrayFilterTupleFunction from "./ArrayFilterTupleFunction";
  * @param filter - Filter function.
  * @returns Array of 2 elements (the matching and the non matching values).
  */
-export const arrayFilterTuple: ArrayFilterTupleFunction = (target, filter) => [
-	arrayFilter(target, filter),
-	arrayFilter(target, (item, index, array) => !filter(item, index, array))
-];
+export const arrayFilterTuple: ArrayFilterTupleFunction = (target, filter) =>
+	objectFreeze([
+		arrayFilter(target, filter),
+		arrayFilter(target, (item, index, array) => !filter(item, index, array))
+	]);
 
 export default arrayFilterTuple;