+++ src/arrayMap/arrayMap.ts
@@ -1,5 +1,6 @@
 import { EMPTY_ARRAY } from "../constants";
 import isArray from "../isArray";
+import objectFreeze from "../objectFreeze";
 import ArrayMapFunction from "./ArrayMapFunction";
 
 /**
@@ -9,7 +10,14 @@ import ArrayMapFunction from "./ArrayMapFunction";
  * @param mapper - Map callback (to be called for every array item).
  * @returns Mapped array.
  */
-export const arrayMap: ArrayMapFunction = (target, mapper) =>
-	isArray(target) ? target.map(mapper) : EMPTY_ARRAY;
+export const arrayMap: ArrayMapFunction = (target, mapper) => {
+	const frozenTarget = isArray(target) ? objectFreeze(target) : EMPTY_ARRAY;
+
+	return objectFreeze(
+		frozenTarget.map((item, index) =>
+			mapper(objectFreeze(item), index, frozenTarget)
+		)
+	);
+};
 
 export default arrayMap;