+++ src/objectFreeze/index.ts
@@ -1,0 +1,2 @@
+export { default, objectFreeze } from "./objectFreeze";
+export { ObjectFreezeFunction } from "./ObjectFreezeFunction";